Removed explicit kernel32 dependencies
Parts of the FAST solution that require some use of the Windows API (mostly OrcaFlex and ServoDyn functionality to interface with DLLs) has been commented out in order to remove any remaining DLL dependencies to the model to get it running in Simulink Real-Time. However, some Intel Fortran and Intel MKL dependencies still require Windows API, so it has not reached this point yet.