re-added in commit ac35094 additions again
I realized commit ac35094 was a crucial commit trapped in the middle of a bunch of DLL removal work, so I manually re-added all the changes made for this commit without anything from before or after it. That's what I get for not abiding by proper versioning practices and just making a new branch.