Added explicit external velocity inputs
Previously, the hybrid model operated with six inputs corresponding to the 6DOF displacements (or forces if using displacement control, though it was decided that we are using force control so that's no longer relevant). After confirmation that the Qualysis system to be used for the physical model can capture velocities in real time, this update allows for these velocities to be inputted directly into the Simulink S-Function instead of being approximated using backwards differencing from the extenral displacements.